backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pipecat-ai[daily,deepgram,silero]>=0.0.40